Refactor TaskCard and TodoListView
